requirements.txt
torch~=2.3.1
Flask~=3.0.3
torchvision~=0.18.1
pillow~=10.4.0
numpy~=2.0.0
pycocotools~=2.0.8